# demos/sample-rest-stockapp/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Documentation",
        "version": "1.0",
        "title": "Api Documentation",
        "termsOfService": "urn:tos",
        "contact": {},
        "license": { "name": "Apache 2.0", "url": "http://www.apache.org/licenses/LICENSE-2.0" }
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "paths": {
...
        "/v1/index": {
            "get": {
                "tags": ["Secondary Index methods"],
                "summary": "Return information about multiple secondary indices.",
                "operationId": "indexInformation",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "namespace",
                        "in": "query",
                        "description": "If specified, the list of returned indices will only contain entries from this namespace.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/RestClientIndex" }
                        }
                    },
                    "403": {
                        "description": "Not authorized to access the resource.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified namespace not found",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": ["Secondary Index methods"],
                "summary": "Create a secondary index.",
                "operationId": "createIndex",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "indexModel",
                        "description": "indexModel",
                        "required": true,
                        "schema": { "$ref": "#/definitions/RestClientIndex" }
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "400": {
                        "description": "Invalid index creation parameters.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to access the resource.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "409": {
                        "description": "Index with the same name already exists, or equivalent index exists.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/index/{namespace}/{name}": {
            "get": {
                "tags": ["Secondary Index methods"],
                "summary": "Get Information about a single secondary index.",
                "operationId": "getIndexStats",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the index",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "The namespace containing the index",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "type": "object", "additionalProperties": { "type": "string" } }
                    },
                    "403": {
                        "description": "Not authorized to access the resource.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified Index does not exist.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            },
            "delete": {
                "tags": ["Secondary Index methods"],
                "summary": "Remove a secondary Index",
                "operationId": "dropIndex",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the index",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "The namespace containing the index",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "403": {
                        "description": "Not authorized to access the resource.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Specified Index does not exist.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
...
        "/v1/truncate/{namespace}": {
            "delete": {
                "tags": ["Truncate Operations"],
                "summary": "Truncate records in a specified namespace.",
                "operationId": "truncateNamespace",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "All records last updated before this date/time will be truncated. If not specified, all records will be truncated.\nThis is a string representation of a date time utilizing the ISO-8601 extended offset date-time format. example: 2019-12-03T10:15:30+01:00",
                        "required": false,
                        "type": "string",
                        "allowEmptyValue": false,
                        "x-example": "2019-12-03T10:15:30+01:00"
                    },
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "The namespace whose records will be truncated.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "400": {
                        "description": "Invalid parameters or request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to access the resource",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/truncate/{namespace}/{set}": {
            "delete": {
                "tags": ["Truncate Operations"],
                "summary": "Truncate records in a specified namespace and set.",
                "operationId": "truncateSet",
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "All records last updated before this date/time will be truncated. If not specified, all records will be truncated.\nThis is a string representation of a date time utilizing the ISO-8601 extended offset date-time format. example: 2019-12-03T10:15:30+01:00",
                        "required": false,
                        "type": "string",
                        "allowEmptyValue": false,
                        "x-example": "2019-12-03T10:15:30+01:00"
                    },
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "The namespace whose records will be truncated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "set",
                        "in": "path",
                        "description": "The set, in the specified namespace, whose records will be truncated",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": { "description": "Accepted" },
                    "400": {
                        "description": "Invalid parameters or request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to access the resource",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        }
    },
    "definitions": {
...
        "RestClientError": {
            "type": "object",
            "properties": {
                "inDoubt": {
                    "type": "boolean",
                    "example": false,
                    "description": "A boolean specifying whether it was possible that the operation succeeded. This is only included if true."
                },
                "message": {
                    "type": "string",
                    "example": "Error Message",
                    "description": "A message describing the cause of the error."
                }
            },
            "title": "RestClientError",
            "description": "Error object returned from unsuccesful operations."
        },
        "RestClientIndex": {
            "type": "object",
            "properties": {
                "bin": {
                    "type": "string",
                    "example": "ageBin",
                    "description": "The bin which is indexed"
                },
                "collection_type": {
                    "type": "string",
                    "enum": ["DEFAULT", "LIST", "MAPKEYS", "MAPVALUES"]
                },
                "name": {
                    "type": "string",
                    "example": "ageIndex",
                    "description": "The name of the index. This must be unique per set"
                },
                "namespace": { "type": "string", "example": "testNS" },
                "set": { "type": "string", "example": "testSet" },
                "type": { "type": "string", "enum": ["NUMERIC", "STRING", "GEO2DSPHERE"] }
            },
            "title": "RestClientIndex"
        },
...
    }
}